Added permissions

+++ src/com/adamki11s/commands/QuestXCommands.java
@@ -6,12 +6,10 @@ import java.util.HashMap;
 
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
-import org.bukkit.Material;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-import org.bukkit.inventory.ItemStack;
 
 import com.adamki11s.ai.RandomMovement;
 import com.adamki11s.display.FixedSpawnsDisplay;
@@ -63,9 +61,7 @@ public class QuestXCommands implements CommandExecutor {
 			} else {
 				Player p = (Player) sender;
 
-				ItemStack[] gear = new ItemStack[] { null, null, null, null, new ItemStack(Material.WOOD_AXE) };
-
-				if (args.length == 1 && args[0].equalsIgnoreCase("force-update")) {
+				if (args.length == 1 && args[0].equalsIgnoreCase("force-update") && p.hasPermission("questsx.admin.update")) {
 					Updater u = new Updater(QuestX.p, "questx", QuestX.f, UpdateType.DEFAULT, true);
 					if (u.getResult() == UpdateResult.SUCCESS) {
 						QuestX.logChat(p, "QuestX version " + u.getLatestVersionString() + " was updated successfully!");
@@ -86,7 +82,7 @@ public class QuestXCommands implements CommandExecutor {
 					return true;
 				}
 
-				if (args.length == 3 && args[0].equalsIgnoreCase("quest") && args[1].equalsIgnoreCase("unpack")) {
+				if (args.length == 3 && args[0].equalsIgnoreCase("quest") && args[1].equalsIgnoreCase("unpack") && p.hasPermission("questsx.quests.setup")) {
 					String qName = args[2];
 					QuestUnpacker upack = new QuestUnpacker(qName);
 					boolean suc = upack.unpackQuest();
@@ -99,7 +95,7 @@ public class QuestXCommands implements CommandExecutor {
 					return true;
 				}
 
-				if (args.length == 3 && args[0].equalsIgnoreCase("quest") && args[1].equalsIgnoreCase("setup")) {
+				if (args.length == 3 && args[0].equalsIgnoreCase("quest") && args[1].equalsIgnoreCase("setup") && p.hasPermission("questsx.quests.setup")) {
 					String qName = args[2];
 					if (setups.containsKey(p.getName())) {
 						QuestX.logChat(p, "You are already setting this quest up!");
@@ -126,7 +122,7 @@ public class QuestXCommands implements CommandExecutor {
 					return true;
 				}
 
-				if (args.length == 2 && args[0].equalsIgnoreCase("quest") && args[1].equalsIgnoreCase("next")) {
+				if (args.length == 2 && args[0].equalsIgnoreCase("quest") && args[1].equalsIgnoreCase("next") && p.hasPermission("questsx.quests.setup")) {
 					if (setups.containsKey(p.getName())) {
 						QuestSetup qs = this.setups.get(p.getName());
 						if (!qs.isSetupComplete()) {
@@ -151,7 +147,7 @@ public class QuestXCommands implements CommandExecutor {
 				 * NPC Commands (START)
 				 */
 
-				if (args.length >= 2 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("list")) {
+				if (args.length >= 2 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("list") && p.hasPermission("questsx.npcs.list")) {
 					String[] list = new String[handle.getNPCs().size()];
 					int count = 0;
 					for (SimpleNPC npc : handle.getNPCs()) {
@@ -196,7 +192,7 @@ public class QuestXCommands implements CommandExecutor {
 					return true;
 				}
 
-				if (args.length == 3 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("delete")) {
+				if (args.length == 3 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("delete") && p.hasPermission("questsx.npcs.delete")) {
 					String toDel = args[2];
 					try {
 						FileUtils.deleteDirectory(FileLocator.getNPCRootDir(toDel));
@@ -212,7 +208,7 @@ public class QuestXCommands implements CommandExecutor {
 					}
 				}
 
-				if (args.length == 3 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("find")) {
+				if (args.length == 3 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("find") && p.hasPermission("questsx.npcs.find")) {
 					String npcName = args[2];
 					SimpleNPC npc = this.handle.getSimpleNPCByName(npcName);
 					if (npc == null) {
@@ -226,7 +222,7 @@ public class QuestXCommands implements CommandExecutor {
 					}
 				}
 
-				if (args.length == 3 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("tele")) {
+				if (args.length == 3 && args[0].equalsIgnoreCase("npc") && args[1].equalsIgnoreCase("tele") && p.hasPermission("questsx.npcs.tele")) {
 					String npcName = args[2];
 					SimpleNPC npc = this.handle.getSimpleNPCByName(npcName);
 					if (npc == null) {
@@ -266,14 +262,14 @@ public class QuestXCommands implements CommandExecutor {
 
 				if (args.length == 3 && args[0].equalsIgnoreCase("fixedspawns")) {
 					String npcName = args[2];
-					if (args[1].equalsIgnoreCase("remove")) {
+					if (args[1].equalsIgnoreCase("delete") && p.hasPermission("questsx.fixedspawns.delete")) {
 						FixedLoadingTable.removeFixedNPCSpawn(p, npcName, handle);
 						return true;
-					} else if (args[1].equalsIgnoreCase("edit")) {
+					} else if (args[1].equalsIgnoreCase("edit") && p.hasPermission("questsx.fixedspawns.edit")) {
 						FixedLoadingTable.editFixedNPCSpawn(p, npcName, handle);
 						return true;
 					} else if (args[1].equalsIgnoreCase("removeall")) {
-						if (p.isOp()) {
+						if (p.hasPermission("questsx.fixedspawns.deleteall")) {
 							FixedLoadingTable.deleteAllFixedSpawns(p, handle);
 						} else {
 							QuestX.logChatError(p, "You must be an Operator to perform this command");
@@ -282,8 +278,8 @@ public class QuestXCommands implements CommandExecutor {
 					}
 				}
 
-				if (args.length >= 2 && args[0].equalsIgnoreCase("fixedspawns")) {
-					if (args[1].equalsIgnoreCase("display")) {
+				if (args.length >= 2 && args[0].equalsIgnoreCase("fixedspawns") && p.hasPermission("questsx.fixedspawns.list")) {
+					if (args[1].equalsIgnoreCase("list")) {
 						if (args.length == 2) {
 							FixedSpawnsDisplay.display(p, 1);
 						} else if (args.length == 3) {
@@ -299,7 +295,7 @@ public class QuestXCommands implements CommandExecutor {
 					}
 				}
 
-				if (args.length == 3 && args[0].equalsIgnoreCase("fixedspawns") && args[1].equalsIgnoreCase("add")) {
+				if (args.length == 3 && args[0].equalsIgnoreCase("fixedspawns") && args[1].equalsIgnoreCase("add") && p.hasPermission("questsx.fixedspawns.add")) {
 					String npcName = args[2];
 					boolean suc = FixedLoadingTable.addFixedNPCSpawn(p, npcName, p.getLocation(), handle);
 					if (suc) {
@@ -319,7 +315,7 @@ public class QuestXCommands implements CommandExecutor {
 				 * Hotspot Commands (START)
 				 */
 
-				if (args.length >= 2 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("list")) {
+				if (args.length >= 2 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("list") && p.hasPermission("questsx.hotspots.list")) {
 					if (args.length == 2) {
 						Pages pages = new Pages(HotspotManager.getAlphabeticalHotspots(), 10);
 						String[] send = pages.getStringsToSend(1);
@@ -352,7 +348,7 @@ public class QuestXCommands implements CommandExecutor {
 					return true;
 				}
 
-				if (args.length == 5 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("add")) {
+				if (args.length == 5 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("add") && p.hasPermission("questsx.hotspots.add")) {
 					String name = args[2];
 
 					if (HotspotManager.doesHotspotExist(name)) {
@@ -376,7 +372,7 @@ public class QuestXCommands implements CommandExecutor {
 					return true;
 				}
 
-				if (args.length == 3 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("delete")) {
+				if (args.length == 3 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("delete") && p.hasPermission("questsx.hotspots.delete")) {
 					String name = args[2];
 					if (HotspotManager.doesHotspotExist(name)) {
 						HotspotManager.deleteHotspot(name);
@@ -387,7 +383,7 @@ public class QuestXCommands implements CommandExecutor {
 					return true;
 				}
 
-				if (args.length == 5 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("edit")) {
+				if (args.length == 5 && args[0].equalsIgnoreCase("hotspots") && args[1].equalsIgnoreCase("edit") && p.hasPermission("questsx.hotspots.edit")) {
 					String name = args[2];
 
 					if (!HotspotManager.doesHotspotExist(name)) {
